+++ packages/lit-virtualizer/src/lit-virtualizer.ts
@@ -1,4 +1,4 @@
 export { repeat } from './lib/repeat';
 export { scroll } from './lib/scroll';
-export { Layout1d, Layout1dGrid } from './lib/uni-virtualizer/uni-virtualizer';
+export { Layout1d, Layout1dGrid, RangeChangeEvent } from './lib/uni-virtualizer/uni-virtualizer';
 export { LitVirtualizer } from './lib/lit-virtualizer';