+++ packages/uni-virtualizer/src/uni-virtualizer.ts
@@ -1,5 +1,5 @@
 export { VirtualRepeater } from './lib/VirtualRepeater';
-export { VirtualScroller } from './lib/VirtualScroller';
+export { VirtualScroller, RangeChangeEvent } from './lib/VirtualScroller';
 
 export { Layout1d } from './lib/layouts/Layout1d';
 export { Layout1dGrid } from './lib/layouts/Layout1dGrid';